+++ src/SMAPI/Constants.cs
@@ -60,7 +60,7 @@ namespace StardewModdingAPI
         internal static int? LogScreenId { get; set; }
 
         /// <summary>SMAPI's current raw semantic version.</summary>
-        internal static string RawApiVersion = "3.12.0";
+        internal static string RawApiVersion = "3.12.1";
     }
 
     /// <summary>Contains SMAPI's constants and assumptions.</summary>